refactor: consolidate SampleView and GenomeSpy into same place
prepare for modularization
