+++ packages/genome-spy/src/app/sampleView/compositeAttributeInfoSource.js
