+++ packages/genome-spy/src/app/sampleView/locations.js
