+++ packages/genome-spy/src/app/sampleView/payloadTypes.d.ts
