+++ packages/genome-spy/src/app/sampleView/sampleState.d.ts
