+++ packages/genome-spy/src/app/sampleView/sampleViewTypes.d.ts
